combine timeseries cache and template device cache
